test(scripts): share scenario fixtures
