+++ test/scripts/sqlite-sessions-transcripts-flip-proof.e2e.test.ts
@@ -1,13 +1,13 @@
 // SQLite sessions/transcripts flip proof test runs the script-style gateway lifecycle probe.
 import { describe, expect, it } from "vitest";
+import { assertSqliteFlipProofCore } from "../helpers/sqlite-sessions-transcripts-flip-proof-assertions.ts";
 import { runSqliteSessionsTranscriptsFlipProof } from "../helpers/sqlite-sessions-transcripts-flip-proof.ts";
 
 describe("SQLite sessions/transcripts flip proof harness", () => {
   it("proves isolated gateway lifecycle state stays SQLite-first", async () => {
     const report = await runSqliteSessionsTranscriptsFlipProof();
 
-    expect(report.failures).toEqual([]);
-    expect(report.ok).toBe(true);
+    assertSqliteFlipProofCore(report);
     expect(report.checkpoints.map((checkpoint) => checkpoint.label)).toEqual([
       "seeded-legacy-store",
       "after-startup-import",
@@ -32,49 +32,9 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
       "after-shared-final-delete",
       "after-final-doctor-inspect",
     ]);
-    expect(
-      report.checkpoints
-        .filter((checkpoint) => checkpoint.label !== "seeded-legacy-store")
-        .every((checkpoint) => checkpoint.activeJsonl.length === 0),
-    ).toBe(true);
-    expect(
-      report.checkpoints.some(
-        (checkpoint) =>
-          checkpoint.label === "seeded-legacy-store" && checkpoint.legacyStateJsonl.length > 0,
-      ),
-    ).toBe(true);
-    expect(
-      report.checkpoints
-        .filter((checkpoint) => checkpoint.label !== "seeded-legacy-store")
-        .every((checkpoint) => checkpoint.legacyStateJsonl.length === 0),
-    ).toBe(true);
-    expect(report.checkpoints.some((checkpoint) => checkpoint.label === "after-doctor-fix")).toBe(
-      false,
-    );
-    expect(
-      report.checkpoints.some(
-        (checkpoint) =>
-          checkpoint.label === "after-startup-import" &&
-          checkpoint.gatewayLogTail?.includes(
-            "session: imported legacy session metadata/transcripts into SQLite",
-          ) === true &&
-          report.oldStateSessionKeys.every((key) =>
-            checkpoint.sqlite.trackedEntries.some((entry) => entry.sessionKey === key),
-          ) &&
-          checkpoint.sqlite.sessionEntries >= 7 &&
-          checkpoint.sqlite.transcriptEvents >= 13,
-      ),
-    ).toBe(true);
     const startupImportCheckpoint = report.checkpoints.find(
       (checkpoint) => checkpoint.label === "after-startup-import",
     );
-    expect(
-      startupImportCheckpoint?.archiveArtifacts.some(
-        (artifact) =>
-          artifact.path.includes(`${report.legacySessionId}.trajectory.jsonl`) &&
-          artifact.textTail?.includes("trajectory") === true,
-      ),
-    ).toBe(true);
     expect(
       startupImportCheckpoint?.archiveArtifacts.some(
         (artifact) =>
@@ -99,15 +59,6 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
         filePath.endsWith("/sqlite-rollback-restore.jsonl"),
       ),
     ).toBe(true);
-    expect(
-      report.checkpoints.some(
-        (checkpoint) =>
-          checkpoint.label === "after-chat-send" &&
-          checkpoint.sqlite.trackedEntries.some(
-            (entry) => entry.sessionKey === report.resetSessionKey && entry.transcriptEvents >= 3,
-          ),
-      ),
-    ).toBe(true);
     expect(report.scaleMigration).toMatchObject({
       minTranscriptEventsPerSession: 4,
       seededEvents: 96,
@@ -115,17 +66,6 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
     });
     expect(report.scaleMigration?.importedSessionKeys).toHaveLength(24);
     expect(report.scaleMigration?.startupImportElapsedMs).toBeGreaterThanOrEqual(0);
-    const resetCheckpoint = report.checkpoints.find(
-      (checkpoint) => checkpoint.label === "after-sessions-reset",
-    );
-    // Retained history: reset keeps the old generation's SQLite rows and
-    // writes no reset archive artifact.
-    const resetArchive = resetCheckpoint?.archiveArtifacts.find(
-      (artifact) =>
-        artifact.archiveReason === "reset" && artifact.archiveSessionId === report.legacySessionId,
-    );
-    expect(resetArchive).toBeUndefined();
-    expect(resetCheckpoint?.sqlite.transcriptEvents ?? 0).toBeGreaterThan(0);
     expect(
       report.checkpoints.some(
         (checkpoint) =>
@@ -160,16 +100,10 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
       ),
     ).toBe(true);
     expect(report.pluginSdkConsumer).toMatchObject({
-      activeJsonlForSessionExists: false,
       activeTrajectoryPointerForSessionExists: false,
       activeTrajectoryRuntimeSidecarForSessionExists: false,
       activeTrajectorySessionSidecarForSessionExists: false,
-      latestAssistantTextBeforeAppend: report.fullTurnAssistantText,
-      latestAssistantTextAfterAppend: "sqlite sdk consumer appended by identity",
-      sessionKey: report.pluginSdkSessionKey,
     });
-    expect(report.pluginSdkConsumer?.sessionFileMarker.startsWith("sqlite:")).toBe(true);
-    expect(report.pluginSdkConsumer?.listedSessionKeys).toContain(report.pluginSdkSessionKey);
     expect(report.pluginSdkConsumer?.transcriptEventsAfterAppend).toBeGreaterThan(
       report.pluginSdkConsumer?.transcriptEventsBeforeAppend ?? 0,
     );
@@ -186,31 +120,6 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
           ),
       ),
     ).toBe(true);
-    const cleanupCheckpoint = report.checkpoints.find(
-      (checkpoint) => checkpoint.label === "after-cleanup-pruning",
-    );
-    expect(
-      cleanupCheckpoint?.sqlite.trackedEntries.some(
-        (entry) => entry.sessionKey === report.cleanupPruneSessionKey,
-      ),
-    ).toBe(false);
-    const cleanupArchive = cleanupCheckpoint?.archiveArtifacts.find(
-      (artifact) =>
-        artifact.archiveReason === "deleted" &&
-        artifact.archiveSessionId === "sqlite-cleanup-prune",
-    );
-    expect(cleanupArchive?.messageTexts).toContain("sqlite cleanup prune me");
-    const idempotenceCheckpoint = report.checkpoints.find(
-      (checkpoint) => checkpoint.label === "after-doctor-import-idempotence",
-    );
-    expect(idempotenceCheckpoint?.doctor).toMatchObject({
-      code: 0,
-      mode: "import",
-      totals: expect.objectContaining({
-        importedEntries: 0,
-        importedTranscriptEvents: 0,
-      }),
-    });
     expect(report.downgradeReupgrade).toMatchObject({
       activeJsonlArchived: true,
       doctorImportedEntries: 1,
@@ -258,24 +167,6 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
       transcriptEvents: 2,
     });
     expect(report.busyContention?.elapsedMs).toBeGreaterThanOrEqual(250);
-    const concurrentCheckpoint = report.checkpoints.find(
-      (checkpoint) => checkpoint.label === "after-concurrent-multi-client",
-    );
-    expect(concurrentCheckpoint).toBeDefined();
-    const concurrentSend = concurrentCheckpoint?.sqlite.trackedEntries.find(
-      (entry) => entry.sessionKey === report.concurrentSendSessionKey,
-    );
-    expect(concurrentSend?.transcriptEvents).toBeGreaterThanOrEqual(2);
-    expect(
-      concurrentCheckpoint?.sqlite.trackedEntries.some(
-        (entry) => entry.sessionKey === report.concurrentResetSessionKey && entry.sessionId,
-      ),
-    ).toBe(true);
-    expect(
-      concurrentCheckpoint?.sqlite.trackedEntries.some(
-        (entry) => entry.sessionKey === report.concurrentDeleteSessionKey,
-      ),
-    ).toBe(false);
     expect(report.secondStartupAfterReset).toMatchObject({
       activeJsonlForSessionExists: false,
       historyContainsPostResetAppend: true,
@@ -300,16 +191,6 @@ describe("SQLite sessions/transcripts flip proof harness", () => {
         artifact.archiveSessionId === "sqlite-delete-session",
     );
     expect(deleteArchive?.messageTexts).toContain("delete me");
-    const sharedFirstCheckpoint = report.checkpoints.find(
-      (checkpoint) => checkpoint.label === "after-shared-first-delete",
-    );
-    expect(
-      sharedFirstCheckpoint?.archiveArtifacts.some(
-        (artifact) =>
-          artifact.archiveReason === "deleted" &&
-          artifact.archiveSessionId === "sqlite-shared-session",
-      ),
-    ).toBe(false);
     const sharedFinalCheckpoint = report.checkpoints.find(
       (checkpoint) => checkpoint.label === "after-shared-final-delete",
     );
